+++ src/BiDaFlow/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `FluentDataflow.LinkWithProbe` to log the behavior of the block
+
 ### Fixed
 - `TransformWithoutBufferBlock` would not work with multiple producers and a slow consumer
 - `FluentDataflow.AsSourceBlock(IEnumerable<T>)` would not dispose the enumerator when the cancel is requested